Split step file
